the returned XYZCT array needs to have an explicit type so that MatlabHandler.pl will know how to add the pixels to OMEIS.
git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@4825 61f9eca8-113b-463c-a1a4-bab9b0a29f14

+++ src/matlab/Attic/Classifier/preprocessing/imgPreProc_band_pass.m
@@ -27,7 +27,7 @@ end
 angle = findOrientationAngle(smaller_img);
 
 % special Nikita sauce to pre-process images
-smaller_img = bandPassIMG(smaller_img);
+smaller_img = uint8(bandPassIMG(smaller_img));
 
 % orient image horizontally (uniform orientation critical to eigenface analysis)
 final_image = imrotate(smaller_img,angle,'bicubic');
